# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_post_ingest_api_v1_ingest_post": {
        "properties": {
          "files": {
            "description": "人员/飞机/课目/规则等文件",
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_post_ingest_api_v1_ingest_post",
        "type": "object"
      },
...
      "ChangeSetView": {
        "additionalProperties": false,
        "description": "`GET /api/v1/ingest/{id}/changeset` 的响应（含 §5.5 冲突项）。",
        "properties": {
          "base_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Snapshot Id"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ChangeItemView"
            },
            "title": "Changes",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictView"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "extraction_ok": {
            "default": true,
            "title": "Extraction Ok",
            "type": "boolean"
          },
          "ingest_job_id": {
            "minLength": 1,
            "title": "Ingest Job Id",
            "type": "string"
          },
          "open_questions": {
            "items": {
              "$ref": "#/components/schemas/OpenQuestionView"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "ingest_job_id"
        ],
        "title": "ChangeSetView",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "description": "`POST /api/v1/chat` —— 统一自然语言入口。幂等键 = 客户端 UUID。",
        "properties": {
          "client_request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "week_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Start"
          }
        },
        "required": [
          "message",
          "client_request_id"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "DecisionRequest": {
        "additionalProperties": false,
        "description": "`approve` / `reject` 的请求体。\n\n`authorized_tiers` 在这里是**真的授权**（v6 §7.2.4）：Planner 那一步只做了\n角色够不够格的预筛，这一步是训练主任本人按下的确认。",
        "properties": {
          "authorized_tiers": {
            "items": {
              "type": "integer"
            },
            "title": "Authorized Tiers",
            "type": "array"
          },
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          }
        },
        "title": "DecisionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestConfirmRequest": {
        "additionalProperties": false,
        "description": "`POST /api/v1/ingest/{id}/confirm`。\n\n`resolutions` / `answers` **不给默认值**：§5.1.1 明令「缺什么问什么，不设\n静默默认值」。冲突没裁完、问题没答完就确认，服务端按 FTS-1004 / FTS-2001 挡回。",
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "问题 id → 回答",
            "title": "Answers",
            "type": "object"
          },
          "approver": {
            "minLength": 1,
            "title": "Approver",
            "type": "string"
          },
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "resolutions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "冲突 id → 选定取值",
            "title": "Resolutions",
            "type": "object"
          }
        },
        "required": [
          "approver"
        ],
        "title": "IngestConfirmRequest",
        "type": "object"
      },
      "IngestConfirmView": {
        "additionalProperties": false,
        "description": "确认后的落库结果。",
        "properties": {
          "applied_resolutions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Applied Resolutions",
            "type": "object"
          },
          "idempotent_hit": {
            "default": false,
            "title": "Idempotent Hit",
            "type": "boolean"
          },
          "snapshot_id": {
            "minLength": 1,
            "title": "Snapshot Id",
            "type": "string"
          },
          "table_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Table Counts",
            "type": "object"
          },
          "vector_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Vector Counts",
            "type": "object"
          }
        },
        "required": [
          "snapshot_id"
        ],
        "title": "IngestConfirmView",
        "type": "object"
      },
      "IngestSubmitView": {
        "additionalProperties": false,
        "description": "`POST /api/v1/ingest` 的响应。幂等键 = 文件 SHA256。",
        "properties": {
          "content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Sha256",
            "type": "string"
          },
          "filenames": {
            "items": {
              "type": "string"
            },
            "title": "Filenames",
            "type": "array"
          },
          "idempotent_hit": {
            "default": false,
            "title": "Idempotent Hit",
            "type": "boolean"
          },
          "ingest_job_id": {
            "minLength": 1,
            "title": "Ingest Job Id",
            "type": "string"
          },
          "trace_id": {
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "ingest_job_id",
          "content_sha256",
          "trace_id"
        ],
        "title": "IngestSubmitView",
        "type": "object"
      },
      "JobStage": {
        "description": "v6 §8.1 的六个粗粒度阶段。**只有六个，不许加第七个。**\n\n加一个阶段就意味着前端的进度条要跟着改，而 v6 §8.1 把「轮询只返回粗粒度\n阶段」当成设计约束（它换来的是几百字节的响应体与可忽略的 rerun 开销）。\n要看细节的地方是 `GET /runs/{trace_id}` 的 `trace_events`。",
        "enum": [
          "解析意图",
          "加载数据",
          "建模",
          "求解",
          "校验",
          "生成报表"
        ],
        "title": "JobStage",
        "type": "string"
      },
      "JobStatus": {
        "description": "运行状态。与阶段正交：`AWAITING_HUMAN` 是「停在某个阶段等人」。",
        "enum": [
          "QUEUED",
          "RUNNING",
          "AWAITING_HUMAN",
          "DONE",
          "FAILED"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusView": {
        "additionalProperties": false,
        "description": "`GET /api/v1/jobs/{job_id}` —— **轮询专用，必须小**（v6 §8.1）。\n\n没有方案、没有校验报告、没有事件明细。要那些就去 `GET /runs/{trace_id}`。",
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent",
            "type": "integer"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trace_id": {
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "trace_id",
          "status",
          "stage",
          "percent"
        ],
        "title": "JobStatusView",
        "type": "object"
      },
      "JobSubmitView": {
        "additionalProperties": false,
        "description": "提交类端点的统一响应：立即返回 `job_id`（v6 §9.2）。",
        "properties": {
          "idempotent_hit": {
            "default": false,
            "title": "Idempotent Hit",
            "type": "boolean"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "poll_url": {
            "minLength": 1,
            "title": "Poll Url",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trace_id": {
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "trace_id",
          "status",
          "poll_url"
        ],
        "title": "JobSubmitView",
        "type": "object"
      },
...
      "ScheduleRequest": {
        "additionalProperties": false,
        "description": "`POST /api/v1/schedule` —— 结构化排班入口。\n\n**它同时是 FTS-4001 的降级路径**（v6 §9.3 脚注：LLM 挂了，排班能力必须还在）。\n所以这里的字段全是结构化的，一个自然语言字段都没有——走这条路不碰 LLM。",
        "properties": {
          "aircraft_ids": {
            "items": {
              "type": "string"
            },
            "title": "Aircraft Ids",
            "type": "array"
          },
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "不给就按请求体内容算幂等键",
            "title": "Client Request Id"
          },
          "person_ids": {
            "items": {
              "type": "string"
            },
            "title": "Person Ids",
            "type": "array"
          },
          "relaxation_tier": {
            "default": 0,
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Relaxation Tier",
            "type": "integer"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "week_start": {
            "description": "排班周周一",
            "format": "date",
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": [
          "week_start"
        ],
        "title": "ScheduleRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "全离线内网部署的飞行训练排班系统。**排班由 CP-SAT 求解、由一套与求解器实现\n完全独立的校验器兜底**；LLM 只做翻译、组织权衡与解释，**不生成、不修改任何一条\n架次记录**（v6 §0.1）。\n\n- 长任务一律异步：提交返回 `job_id`，轮询 `GET /jobs/{job_id}`（阶段 + 百分比），\n  完成后一次性取 `GET /runs/{trace_id}`（方案 + 校验报告 + TraceEvent 全量）。\n- 三态分离：`OPTIMAL/FEASIBLE` / `INFEASIBLE` / `UNKNOWN` 在错误码与数据模型上\n  全程不混（铁律 8）。\n- 认证：`Authorization: Bearer <token>`，角色 viewer < scheduler < director < admin。\n",
    "title": "FTS 飞行训练排班系统 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/chat": {
      "post": {
        "description": "提交一句话，立即返回 `job_id` + `trace_id`（v6 §8.1）。\n\n`thread_id` 非空时**续接**那次运行（修订轮走这条路）：trace_id 沿用，\nLangGraph 按同一个 thread 从 checkpoint 接着跑。",
        "operationId": "post_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmitView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "自然语言入口（异步，返回 job_id）",
        "tags": [
          "会话"
        ]
      }
    },
    "/api/v1/ingest": {
      "post": {
        "description": "存盘 + 登记。**不在这里解析**——解析在 `/changeset`。\n\n多文件时的幂等键是「每份文件摘要按文件名排序后再摘要」：一批文件的身份\n应当与上传顺序无关，而 `sha256(a||b) != sha256(b||a)`。",
        "operationId": "post_ingest_api_v1_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_ingest_api_v1_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSubmitView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "上传数据文件（幂等键 = 文件 SHA256）",
        "tags": [
          "摄取"
        ]
      }
    },
    "/api/v1/ingest/{job_id}/changeset": {
      "get": {
        "operationId": "get_changeset_api_v1_ingest__job_id__changeset_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取待确认 Diff（含 §5.5 的冲突项与待答问题）",
        "tags": [
          "摄取"
        ]
      }
    },
    "/api/v1/ingest/{job_id}/confirm": {
      "post": {
        "operationId": "post_confirm_api_v1_ingest__job_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestConfirmView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "确认变更 → 落库 → 生成新 snapshot",
        "tags": [
          "摄取"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "轮询任务状态（阶段 + 百分比 + 状态）",
        "tags": [
          "任务"
        ]
      }
    },
...
    "/api/v1/runs/{trace_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取回完整结果：方案 + 校验报告 + TraceEvent 全量",
        "tags": [
          "任务"
        ]
      }
    },
    "/api/v1/schedule": {
      "post": {
        "operationId": "post_schedule_api_v1_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmitView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "结构化排班入口（异步，零 LLM 依赖）",
        "tags": [
          "排班"
        ]
      }
    },
    "/api/v1/schedule/{trace_id}/approve": {
      "post": {
        "operationId": "post_approve_api_v1_schedule__trace_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "人工确认 → 归档 + 回写记忆（需训练主任）",
        "tags": [
          "排班"
        ]
      }
    },
    "/api/v1/schedule/{trace_id}/export": {
      "get": {
        "description": "取归档产物。**只在归档之后才有文件**（`commit_plan` 写的那个）。\n\n没有文件时报 FTS-1004 而不是 404 空响应：用户要的下一步很明确\n——先去人工门禁确认，产物才会生成。",
        "operationId": "get_export_api_v1_schedule__trace_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "下载本次运行归档的 xlsx",
        "tags": [
          "排班"
        ]
      }
    },
    "/api/v1/schedule/{trace_id}/reject": {
      "post": {
        "description": "驳回。**不自动重排**——驳回的理由要人来定（v6 §7.2.4 那张表）。",
        "operationId": "post_reject_api_v1_schedule__trace_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "驳回附意见（不自动重排）",
        "tags": [
          "排班"
        ]
      }
    },
...
  }
}